Adding random color tensorflow graphs
Originally, we were training on random size and brightness changes,
but in order to handle any skin pigmentation and background color, we now
run all of the training images through a colorizer filter.

Basically, it just chooses a random shift between 0 and 255 and shifts the
hue of every color in the image by that offset. This gives us an incredibly
diverse sample set and directs the graph to concentrate more on shapes than
colors and positions.